all_twfe_weights

a function to compute implicit TWFE weights

all_twfe_weights(
  yname,
  tname,
  idname,
  gname,
  xformla = NULL,
  xname = NULL,
  data
)

Arguments

yname

The name of the outcome variable

tname

The name of the column containing the time periods

idname

The individual (cross-sectional unit) id name

gname

The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It defines which "group" a unit belongs to. It should be 0 for units in the untreated group.

xformla

formula for covariates that are included in the TWFE regression.

xname

by default, twfe_weights_gt applies the implicit TWFE regression weights to the outcome. In cases where the researcher would like to apply the implicit regression weights to some other variable, the name of this variable can be passed in through this argument.

This is typically used as a way to check how well the TWFE regression balances levels of time-varying covariates, time-invariant covariates, or other variables not included in the model. This is often useful as a diagnostic to assess how sensitive the TWFE regression is to violations of linearity.

@return a gt_weights object @export

data

The name of the data.frame that contains the data

Value

a decomposed_twfe object