gt_weights

a class for holding and working with group-time weights

gt_weights(
  g,
  tp,
  treated = NULL,
  comparison = NULL,
  weights_treated,
  weights_comparison,
  weighted_outcome_treated = NULL,
  weighted_outcome_comparison = NULL,
  weighted_outcome_diff = NULL,
  alpha_weight
)

Arguments

g

the current group for which the weights will be computed

tp

the time period for which the weights will be computed

treated

a vector of units that make up the effective treated group for group g and period tp

comparison

a vector of units that make up the effective comparison group for group g and period tp

weights_treated

a vector of weights for the treated group g in time period tp

weights_comparison

a vector of weights for the comparison group for group g in time period tp

weighted_outcome_treated

the mean outcome for the treated group after applying weights_treated to it

weighted_outcome_comparison

the mean outcome for the comparison group after applying weights_comparison to it

weighted_outcome_diff

the difference between the mean outcomes for the treated and comparison groups after applying weights to them

alpha_weight

the contribution of this term to the estimate of alpha